Fix Campaign load after going back to Campaign index
The `goToCampaign` function dispatchs a navigation event, which in
turn dispatchs a `CAMPAIGN_FETCH` event that clears `state.campaign`
But the `CAMPAIGN_FETCH` event is processed _after_ rendering the
`Campaign` control.

So when the user is at the index and navigates to a campaign, the
`render` method sees that `state.campaign.data` is `null`, and
renders an empty page. Then the `CAMPAIGN_FETCH` and
`CAMPAIGN_FETCHED` resolve, correctly rendering everything.

But if the user hits the browser's back button, `state.campaign.data`
is still populated - and doesn't get nulled. When the user clicks
the next Campaign to see, it navigates, and the `render` method sees
`state.campaign.data` with the old data - and then it crashes.

This commit clears `state.campaign.data` before navigating, so the
issue is avoided at all.

Fixes #63